Legal updates from CELA (#1765)
* This license file is being replaced.

* Update from CELA.

CELA has provided replacement licence files which include separate licenses for CC and code snippets. The Content Engineering Team has been tasked with replacing the license files on all existing repos. The changes to the ReadMe file must be at the top of the file to be recognized by license scanners.

* New license files from CELA.

* Removing the MSFT CoC. Fixing the H1.